requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
pyzbar==0.1.9
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
